requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.35.0
blinker==1.9.0
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
Flask==3.1.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpickle==1.4.2
MarkupSafe==3.0.2
msal==1.32.3
msrest==0.7.1
multidict==6.6.3
numpy==2.2.6
oauthlib==3.3.1
openai==1.95.1
packaging==25.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
